cdk-tgw-outbound-aggregation/requirements.txt
aws-cdk.core==1.27.0
aws-cdk.aws-ec2==1.27.0
aws-cdk.aws-iam==1.27.0
